Break examples into seperate demo components
Add links to source code for each component and example.
