+++ lib/src/demo_app/demo_app.css
@@ -4,12 +4,7 @@
  * BSD-style license that can be found in the LICENSE file.
  */
 
-.source-link:link, .source-link:visited, .source-link:active {
-  display: block;
-  padding-bottom: 10px;
-}
-
-.source-link:hover {
-  display: block;
-  padding-bottom: 10px;
+.blue {
+  background-color: #2196F3;
+  color: white;
 }
