+++ lib/src/demo_app/demo_app.html
@@ -1,0 +1,26 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<p>
+  <material-checkbox
+      [(checked)]="allowed"
+      label="Allow clicking the two buttons below">
+  </material-checkbox>
+</p>
+
+<p>
+  Count: {{count}} &nbsp;&nbsp;&nbsp;
+  <material-button
+      [disabled]="!allowed"
+      (trigger)="increment()"
+      raised class="blue">
+    Increase count
+  </material-button>
+  <material-button
+      [disabled]="!allowed"
+      (trigger)="reset()">
+    Reset
+  </material-button>
+</p>
