+++ lib/src/glyph_demo/glyph_demo.dart
@@ -1,0 +1,15 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:angular2/angular2.dart';
+import 'package:angular_components/angular_components.dart';
+
+@Component(
+  selector: 'glyph-demo',
+  templateUrl: 'glyph_demo.html',
+  directives: const [GlyphComponent],
+)
+class GlyphDemoComponent {
+  // Nothing here.
+}
