+++ lib/src/glyph_demo/glyph_demo.html
@@ -1,0 +1,11 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<glyph icon="favorite"></glyph>
+<glyph icon="business"></glyph>
+<glyph icon="thumb_up"></glyph>
+<glyph icon="bluetooth_connected"></glyph>
+<glyph icon="insert_photo"></glyph>
+<glyph icon="more_horiz"></glyph>
