+++ lib/src/material_button_demo/material_button_demo.css
@@ -4,12 +4,16 @@
  * BSD-style license that can be found in the LICENSE file.
  */
 
-.source-link:link, .source-link:visited, .source-link:active {
-  display: block;
-  padding-bottom: 10px;
+.red:not([raised]) {
+  color: #f44336;
 }
 
-.source-link:hover {
-  display: block;
-  padding-bottom: 10px;
+.red[raised] {
+  background-color: #f44336;
+  color: white;
+}
+
+.white {
+  background-color: white;
+  color: #4285f4;
 }
