+++ lib/src/material_button_demo/material_button_demo.dart
@@ -1,0 +1,20 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:angular2/angular2.dart';
+import 'package:angular_components/angular_components.dart';
+
+@Component(
+  selector: 'material-button-demo',
+  styleUrls: const ['material_button_demo.css'],
+  templateUrl: 'material_button_demo.html',
+  directives: const [
+    MaterialButtonComponent,
+    MaterialFabComponent,
+    GlyphComponent
+  ],
+)
+class MaterialButtonDemoComponent {
+  // Nothing here.
+}
