+++ lib/src/material_button_demo/material_button_demo.html
@@ -1,0 +1,51 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<section class="margin">
+  <h3>Flat</h3>
+  <material-button>button</material-button>
+  <material-button class="red">Colored</material-button>
+  <material-button disabled>Disabled</material-button>
+  <material-button dense>Condensed</material-button>
+</section>
+
+<section class="margin">
+  <h3>Raised</h3>
+  <material-button raised>Raised</material-button>
+  <material-button raised class="red">Colored</material-button>
+  <material-button raised class="red" disabled>Disabled</material-button>
+  <material-button raised dense>Condensed</material-button>
+</section>
+
+<section class="margin">
+  <h3>Without min-width and margin</h3>
+  <p>Regular</p>
+  <material-button raised class="red">button</material-button>
+  <material-button raised class="red">A</material-button>
+  <p>Without min-width and margin</p>
+  <material-button raised clear-size class="red">button</material-button>
+  <material-button raised clear-size class="red">A</material-button>
+</section>
+
+<section class="margin">
+  <h3>With Icon</h3>
+  <material-button icon>
+    <glyph icon="check"></glyph>
+  </material-button>
+</section>
+
+<section class="margin">
+  <h3>
+    <strong>F</strong>loating
+    <strong>A</strong>ction
+    <strong>B</strong>utton
+  </h3>
+  <material-fab raised>
+    <glyph icon="add"></glyph>
+  </material-fab>
+  <material-fab mini raised>
+    <glyph icon="check"></glyph>
+  </material-fab>
+</section>
