+++ lib/src/material_chips_demo/material_chips_demo.html
@@ -1,0 +1,14 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<material-chips>
+  <material-chip>My Chip</material-chip>
+  <material-chip>Another Chip</material-chip>
+</material-chips>
+
+<p>
+  Note: These are not backed by any model on this page, so clicking them won't
+  do anything.
+</p>
