+++ lib/src/material_dialog_demo/material_dialog_demo.css
@@ -1,0 +1,53 @@
+/*
+ * Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+ * for details. All rights reserved. Use of this source code is governed by a
+ * BSD-style license that can be found in the LICENSE file.
+ */
+
+.blue:not([disabled]) {
+  background-color: #2196F3;
+  color: white;
+}
+
+.white:not([disabled]) {
+  background-color: white;
+  color: #4285f4;
+}
+
+.error-button {
+  margin: 8px;
+}
+
+.basic-dialog,
+.basic-scrolling-dialog,
+.max-height-dialog,
+.headered-dialog,
+.custom-colors-dialog,
+.no-header-footer-dialog {
+  width: 480px;
+}
+
+.basic-scrolling-dialog {
+  height: 320px;
+}
+
+.basic-scrolling-dialog > div {
+  padding: 8px;
+}
+
+.max-height-dialog {
+  max-height: 40vh;
+}
+
+.dialog-with-error,
+.info-dialog {
+  width: 320px;
+}
+
+.custom-colors-dialog {
+  background-color: #b7e1cd;
+}
+
+.no-header-footer-dialog {
+  height: 6em;
+}
