+++ lib/src/material_expansionpanel_demo/material_expansionpanel_demo.html
@@ -1,0 +1,17 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<material-expansionpanel-set>
+  <material-expansionpanel name="Expansion panel">
+    <div>
+      Oh hi. I was just trying not to take too much space here.
+    </div>
+  </material-expansionpanel>
+  <material-expansionpanel name="Expansion panel #2">
+    <div>
+      Me too! Don't mind me.
+    </div>
+  </material-expansionpanel>
+</material-expansionpanel-set>
