+++ lib/src/material_input_demo/material_input_demo.html
@@ -1,0 +1,27 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<material-input label="Write something"></material-input>
+<material-input floatingLabel label="Floating label"></material-input>
+<material-input required label="Required"></material-input>
+<material-input [disabled]="true" label="Disabled"></material-input>
+<material-input [maxCount]="5" label="Max 5 chars"></material-input>
+<material-input [rightAlign]="true"
+                label="Aligned to the right"></material-input>
+<material-input leadingText="$" label="With leading text"></material-input>
+<material-input trailingText="USD" [rightAlign]="true"
+                label="With trailing text"></material-input>
+<material-input leadingGlyph="business"
+                label="With leading glyph"></material-input>
+<material-input trailingGlyph="favorite" [rightAlign]="true"
+                label="With trailing glyph"></material-input>
+
+<div>
+  Aligned with
+  <material-input type="text" label="surrounding"></material-input>
+  text.
+</div>
+
+<material-input multiline rows="3" label="Write many lines"></material-input>
