+++ lib/src/material_list_demo/material_list_demo.html
@@ -1,0 +1,141 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<span class="colorchanger" [style.color]="selectedColor">Use the lists to select a color.</span>
+
+<h3>Basic List</h3>
+<material-list>
+  <div group>
+    <material-list-item (trigger)="selectColor('red')">
+      Red
+    </material-list-item>
+    <material-list-item (trigger)="selectColor('green')">
+      Green
+    </material-list-item>
+  </div>
+  <material-list-item (trigger)="selectColor('coral')" disabled>
+    Coral
+  </material-list-item>
+  <material-list-item (trigger)="selectColor('mediumOrchid')">
+    MediumOrchid
+  </material-list-item>
+</material-list>
+
+
+<h3>Different List Sizes</h3>
+<p><em>With borders to demonstrate container size</em></p>
+<p>Available Sizes: auto, x-small, small, medium, large, x-large</p>
+<div *ngFor="let size of sizes">
+  <material-list class="bordered-list"
+                 [attr.size]="size">
+    <div label>List Width: {{size}}</div>
+    <material-list-item (trigger)="selectColor('red')">
+      Red
+    </material-list-item>
+    <material-list-item (trigger)="selectColor('green')">
+      Green
+    </material-list-item>
+    <material-list-item (trigger)="selectColor('coral')">
+      Coral
+    </material-list-item>
+    <material-list-item (trigger)="selectColor('mediumOrchid')">
+      MediumOrchid
+    </material-list-item>
+  </material-list>
+</div>
+
+<h3>Minimum Width</h3>
+<p><em>With borders to demonstrate container width</em></p>
+<material-list class="bordered-list"
+               min-size="x-small">
+  <div label>Min Width: x-small</div>
+  <material-list-item (trigger)="selectColor('red')">
+    Red
+  </material-list-item>
+  <material-list-item (trigger)="selectColor('green')">
+    Green
+  </material-list-item>
+  <material-list-item (trigger)="selectColor('coral')">
+    Coral
+  </material-list-item>
+  <material-list-item (trigger)="selectColor('mediumOrchid')">
+    MediumOrchid
+  </material-list-item>
+</material-list>
+
+<h3>Navigate Focus via Keyboard</h3>
+<p><em>
+  Use &lt;tab&gt; to focus an item in the list, then use up/down and
+  left/right keys to navigate.
+</em></p>
+<material-list focusList tabindex="0" class="bordered-list">
+  <div group>
+    <material-list-item (trigger)="selectColor('red')" focusItem>
+      Red
+    </material-list-item>
+    <material-list-item (trigger)="selectColor('green')" focusItem>
+      Green
+    </material-list-item>
+  </div>
+  <material-list-item (trigger)="selectColor('coral')" focusItem>
+    Coral
+  </material-list-item>
+  <material-list-item (trigger)="selectColor('mediumOrchid')" focusItem>
+    MediumOrchid
+  </material-list-item>
+</material-list>
+
+<h3>Primary and Secondary Item Content</h3>
+<material-list class="bordered-list" width="3">
+  <material-list-item>
+    Item with secondary
+    <span class="material-list-item-secondary">
+      ctrl+x
+    </span>
+  </material-list-item>
+  <material-list-item>
+    <glyph icon="today" class="material-list-item-primary"
+           aria-hidden="true"></glyph>
+    Item with primary icon
+  </material-list-item>
+  <material-list-item>
+    Item
+  </material-list-item>
+  <material-list-item>
+    Item
+  </material-list-item>
+</material-list>
+
+<h3>Grouping and Dividers</h3>
+<material-list class="bordered-list" width="3">
+  <div group>
+    <material-list-item>
+      Item
+    </material-list-item>
+    <material-list-item>
+      Item
+    </material-list-item>
+    <material-list-item>
+      Item with secondary
+      <span class="material-list-item-secondary">
+        ctrl+x
+      </span>
+    </material-list-item>
+  </div>
+  <div group>
+    <material-list-item>
+      Small Group Item
+    </material-list-item>
+  </div>
+  <material-list-item>
+    <glyph icon="today" class="material-list-item-primary"
+           aria-hidden="true"></glyph>
+    Item with primary icon
+  </material-list-item>
+  <material-list-item>
+    Item
+  </material-list-item>
+  <material-list-item>Item</material-list-item>
+</material-list>
