+++ lib/src/material_popup_demo/material_popup_demo.css
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+ * for details. All rights reserved. Use of this source code is governed by a
+ * BSD-style license that can be found in the LICENSE file.
+ */
+
+.blue {
+  background-color: #2196F3;
+  color: white;
+}
+
+.basic {
+  height: 200px;
+  padding: 5px;
+}
+
+.custom-header {
+  background: #43A047;
+  color: white;
+  padding: 5px;
+}
+
+.custom-body {
+  background: #E8F5E9;
+  height: 1000px;
+  padding: 5px;
+  width: 200px;
+}
+
+.custom-footer {
+  background: #1B5E20;
+  color: white;
+  padding: 5px;
+}
+
+.limited-size {
+  background: #EEEEEE;
+  height: 1000px;
+  padding: 5px;
+  width: 1000px;
+}
