+++ lib/src/material_popup_demo/material_popup_demo.html
@@ -1,0 +1,69 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<h3>Basic Popup</h3>
+<material-button class="blue"
+                 raised
+                 popupSource
+                 #source="popupSource"
+                 (trigger)="basicPopupVisible = !basicPopupVisible">
+  {{basicPopupVisible ? 'Close' : 'Open'}} Popup
+</material-button>
+<material-popup [source]="source" [(visible)]="basicPopupVisible">
+  <div class="basic">
+    Hello, I am a pop up!
+  </div>
+</material-popup>
+
+<h3>Headers and Footers</h3>
+<p>
+  A popup with a header and a footer.
+</p>
+<material-button class="blue"
+                 raised
+                 popupSource
+                 #headerExampleSource="popupSource"
+                 (trigger)="headerFooterPopupVisible = !headerFooterPopupVisible">
+  {{headerFooterPopupVisible ? 'Close' : 'Open'}} Custom Popup
+</material-button>
+<material-popup defaultPopupSizeProvider
+                enforceSpaceConstraints
+                [source]="headerExampleSource"
+                [(visible)]="headerFooterPopupVisible">
+  <div header class="custom-header">
+    This is a Header demo
+  </div>
+  <div class="custom-body">
+    Hello, Hello, Hello. This is a tall bit of content that needs a scroll
+    bar because the content is so long.
+  </div>
+  <div footer class="custom-footer">
+    This is a Footer demo
+  </div>
+</material-popup>
+
+<h3>Limited Size</h3>
+<p>
+  Using the default <em>PercentagePopupSizeProvider</em> to limit max width
+  to be 70% of the viewport width and max height to be 50% of the viewport
+  height.
+</p>
+<material-button class="blue"
+                 raised
+                 popupSource
+                 #mixinExampleSource="popupSource"
+                 (trigger)="sizePopupVisible = !sizePopupVisible">
+  {{sizePopupVisible ? 'Close' : 'Open'}} Provider Popup
+</material-button>
+<material-popup defaultPopupSizeProvider
+                enforceSpaceConstraints
+                [source]="mixinExampleSource"
+                [(visible)]="sizePopupVisible">
+  <div class="limited-size">
+    Hello, Hello, Hello. This is a long content that may need a scroll bar
+    if the size of the popup is too wide for the window. We cap it at 50% of
+    the viewport height as a reasonable default.
+  </div>
+</material-popup>
