+++ lib/src/material_progress_demo/material_progress_demo.html
@@ -1,0 +1,9 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<material-progress [activeProgress]="10"
+                   [secondaryProgress]="30"></material-progress>
+
+<material-progress [indeterminate]="true"></material-progress>
