+++ lib/src/material_radio_demo/material_radio_demo.html
@@ -1,0 +1,13 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<material-radio-group>
+  <material-radio [checked]="true">
+    default choice
+  </material-radio>
+  <material-radio>
+    alternative choice
+  </material-radio>
+</material-radio-group>
