+++ lib/src/material_select_demo/material_select_demo.css
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+ * for details. All rights reserved. Use of this source code is governed by a
+ * BSD-style license that can be found in the LICENSE file.
+ */
+
+.bordered-list {
+  border: 1px solid #000;
+  border-radius: 8px;
+  display: inline-block;
+  margin: 8px;
+  overflow: hidden;
+}
+
+.example-with-options {
+  display: flex;
+}
+
+.example-pane {
+  flex: 1;
+}
+
+.options-pane {
+  flex: 1;
+  padding: 16px;
+  background: #eee;
+}
+
+.options-pane > material-checkbox {
+  margin-left: 0;
+}
+
+.option-label {
+  display: inline-flex;
+  padding: 10px 0px;
+  width: 50%;
+}
