+++ lib/src/material_spinner_demo/material_spinner_demo.html
@@ -1,0 +1,6 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<material-spinner></material-spinner> &nbsp; waiting for Godot
