+++ lib/src/material_tab_demo/material_tab_demo.html
@@ -1,0 +1,28 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<material-tab-panel>
+  <material-tab label="tab 1">
+    <div>
+      These are the contents of Tab 1.
+    </div>
+  </material-tab>
+  <material-tab label="tab 2">
+    <div>
+      Tab 2 contents, on the other hand, look thusly.
+    </div>
+  </material-tab>
+  <material-tab label="tab 3">
+    <div>
+      <h3>Tab 3 is serious about its contents</h3>
+      <p>
+        Lorem ipsum dolor sit amet, consectetur adipisicing elit. Labore magni
+        necessitatibus quam qui quis rerum sit, sunt voluptatum. Commodi,
+        corporis minus nemo officiis quisquam rem. Magni odit quo temporibus
+        veritatis!
+      </p>
+    </div>
+  </material-tab>
+</material-tab-panel>
