+++ lib/src/material_toggle_demo/material_toggle_demo.html
@@ -1,0 +1,15 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<material-toggle label="Unchecked"></material-toggle>
+<br/>
+<material-toggle [checked]="true" label="Checked"></material-toggle>
+<br/>
+<material-toggle [disabled]="true" label="Disabled off"></material-toggle>
+<br/>
+<material-toggle [disabled]="true"
+                 [checked]="true"
+                 label="Disabled on">
+</material-toggle>
