+++ lib/src/reorder_list_demo/reorder_list_demo.html
@@ -1,0 +1,12 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<ol>
+  <reorder-list (reorder)="onReorder($event)">
+    <li *ngFor="let item of reorderItems" reorderItem>
+      {{item}}
+    </li>
+  </reorder-list>
+</ol>
