+++ lib/src/scorecard_demo/scorecard_demo.html
@@ -1,0 +1,30 @@
+<!--
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+  for details. All rights reserved. Use of this source code is governed by a
+  BSD-style license that can be found in the LICENSE file.
+-->
+<acx-scorecard
+    selectable
+    label="Estimated earnings"
+    value="$158.22"
+    description="+$24.20 (15%)"
+    suggestionAfter="of last quarter"
+    changeType="POSITIVE">
+</acx-scorecard>
+
+<acx-scorecard
+    selectable
+    label="Lines of code"
+    value="123 KLOC"
+    description="+3535 (2%)"
+    suggestionAfter="of last quarter"
+    changeType="NEGATIVE">
+</acx-scorecard>
+
+<acx-scorecard
+    selectable
+    [selected]="true"
+    label="Selected scorecard"
+    value="9,000"
+    description="Try clicking this">
+</acx-scorecard>
