+++ web/index.html
@@ -1,5 +1,5 @@
 <!--
-  Copyright (c) 2016, the Dart project authors.  Please see the AUTHORS file
+  Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
   for details. All rights reserved. Use of this source code is governed by a
   BSD-style license that can be found in the LICENSE file.
 -->
@@ -30,6 +30,16 @@
         margin: 40px auto;
       }
 
+      a:link, a:visited, a:active {
+        color: rgb(66, 133, 244);
+        text-decoration: none;
+      }
+
+      a:hover {
+        color: rgb(46, 113, 224);
+        text-decoration: none;
+      }
+
       hr {
         border: 0;
         height: 0;
@@ -48,7 +58,7 @@
     </style>
   </head>
   <body>
-    <h1>AngularDart Components Example</h1>
+    <h1>AngularDart Components</h1>
     <p>
       Useful links:
     </p>
@@ -68,7 +78,6 @@
         for AngularDart Components
       </li>
     </ul>
-    <hr>
     <my-app>Loading...</my-app>
   </body>
 </html>
